handle cases when writing to wrtite protected template

+++ system/templates/deployment/global/home/config_functions.sh
@@ -18,18 +18,26 @@ templates=`find /home/engines/templates/ -type f |grep -v keep_me`
                 then
                 	dest_file=`ls -l $dest_file |cut -f2 -d">"`
                 fi
-                
-              #  /home/engines/scripts/grant_rw_access.sh `dirname $dest_file`
-               # /home/engines/scripts/grant_rw_access.sh $dest_file
+                 
+               write=`stat -c  %A LICENSE | awk -F\-  '{ print $3}' |grep w |wc -c`
+               
+               if test write -eq 0
+                then
+                  /home/engines/scripts/grant_rw_access.sh `dirname $dest_file`
+                  /home/engines/scripts/grant_rw_access.sh $dest_file
+                fi
+               
 				rm $dest_file
 				
 				echo doing $dest_file
 				
        			process_file 
        			#FIXME only revoke if it was g+w before
-       			#/home/engines/scripts/revoke_rw_access.sh $dest_file
-       			#/home/engines/scripts/revoke_rw_access.sh `dirname $dest_file`
-
+       			if test write -eq 0
+       			then
+       				/home/engines/scripts/revoke_rw_access.sh $dest_file
+       				/home/engines/scripts/revoke_rw_access.sh `dirname $dest_file`
+				fi
         done
         echo run as `whoami` in `pwd`
 }
